requirements.txt
Flask
Flask-cors
python-dotenv
 